Fix missing UI state models and Kotlin build errors

+++ app/src/main/java/dev/garnetforge/app/data/model/UiStates.kt
@@ -1,31 +1,31 @@
 package dev.garnetforge.app.data.model
 
 sealed interface SpeedTestState {
-    val downloadMbps: Double
-    val uploadMbps: Double
+    val downloadMbps: Float
+    val uploadMbps: Float
     val msg: String
 
     data object Idle : SpeedTestState {
-        override val downloadMbps = 0.0
-        override val uploadMbps = 0.0
+        override val downloadMbps = 0f
+        override val uploadMbps = 0f
         override val msg = ""
     }
 
     data object Running : SpeedTestState {
-        override val downloadMbps = 0.0
-        override val uploadMbps = 0.0
+        override val downloadMbps = 0f
+        override val uploadMbps = 0f
         override val msg = "Running"
     }
 
     data class Done(
-        override val downloadMbps: Double = 0.0,
-        override val uploadMbps: Double = 0.0,
+        override val downloadMbps: Float = 0f,
+        override val uploadMbps: Float = 0f,
         override val msg: String = "Done",
     ) : SpeedTestState
 
     data class Error(override val msg: String) : SpeedTestState {
-        override val downloadMbps = 0.0
-        override val uploadMbps = 0.0
+        override val downloadMbps = 0f
+        override val uploadMbps = 0f
     }
 }
